Check if there is an Onion Service before removing it
It is possible to not have an Onion Service when on local mode, but
also when unMessage fails to add one and does not crash.

+++ unmessage/peer.py
@@ -656,21 +656,22 @@ class Peer(object):
         return d_tor
 
     def _stop_tor(self):
-        self._ui.notify(
-            notifications.UnmessageNotification(
-                'Removing Onion Service from Tor'))
-
-        e = Event()
-
-        def removed(result):
+        if self._onion_service:
             self._ui.notify(
                 notifications.UnmessageNotification(
-                    'Removed Onion Service from Tor'))
-            e.set()
+                    'Removing Onion Service from Tor'))
 
-        d = self._onion_service.remove_from_tor(self._tor._protocol)
-        d.addCallback(removed)
-        e.wait()
+            e = Event()
+
+            def removed(result):
+                self._ui.notify(
+                    notifications.UnmessageNotification(
+                        'Removed Onion Service from Tor'))
+                e.set()
+
+            d = self._onion_service.remove_from_tor(self._tor._protocol)
+            d.addCallback(removed)
+            e.wait()
 
     def _send_request(self, identity, key):
         result = re.match(r'[^@]+@[^:]+(:(\d+))?$', identity)
